Phase-2/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
pydantic-extra-types
email-validator
python-dotenv
psycopg2-binary
requests